rename snappy_ssz to ssz_snappy to better align with wire name
